[plugin.video.playonbrowser] 2.0.4 (#1956)
